extract common method
